requirements.txt
flask
groq
python-dotenv
pypdf
scikit-learn
